pyproject.toml
[tool.poetry]
name = "habitladder"
version = "0.1.0"
description = "HabitLadder is the backend of a SPA web application - a tracker of healthy habits"
authors = ["Nadezhda_Popova <nadezhdapopova13@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.8"
djangorestframework = "^3.16.1"
python-dotenv = "^1.2.1"
psycopg2-binary = "^2.9.11"
django-filter = "^25.2"
djangorestframework-simplejwt = "^5.5.1"
drf-yasg = "^1.21.11"
django-cors-headers = "^4.9.0"
redis = "^7.1.0"
pillow = "^12.0.0"
django-phonenumber-field = "^8.4.0"
phonenumbers = "^9.0.20"
django-cities-light = "^3.10.2"
celery = "^5.6.0"
eventlet = "^0.40.4"
django-celery-beat = "^2.8.1"
celery-types = "^0.23.0"


[tool.poetry.group.dev.dependencies]
poetry-core = "^2.2.1"
setuptools = "^80.9.0"
coverage = "^7.12.0"


[tool.poetry.group.lint.dependencies]
black = "^25.11.0"
flake8 = "^7.3.0"
isort = "^7.0.0"

[tool.black]
line-length = 119
exclude = """
/(
    .git
  | .hg
  | .mypy_cache
  | .tox
  | .venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | static
  | media
  | node_modules
  | __pycache__
  | manage.py
  | settings.py
  | asgi.py
  | wsgi.py
)/
"""

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
skip_gitignore = true
known_django = ["django"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
